Force revalidation of fetched resources
When making a fetch call that does not need to be quick, such as a deferred
code update, the `{ cache: "no-cache" }` flag is supplied to the `fetch`
call.

The `no-cache` directive does actually use the cache, but force validates;
a conditional request that returns a 304 Not Modified will use the cached
value, otherwise the response is downloaded and returned.